Move org-vault folder to vault
