+++ apps/web/src/app/organizations/organization-routing.module.ts
@@ -12,13 +12,14 @@ import {
 } from "@bitwarden/common/abstractions/organization/organization.service.abstraction";
 import { Organization } from "@bitwarden/common/models/domain/organization";
 
+import { VaultModule } from "../../vault/app/org-vault/vault.module";
+
 import { OrganizationPermissionsGuard } from "./guards/org-permissions.guard";
 import { OrganizationRedirectGuard } from "./guards/org-redirect.guard";
 import { OrganizationLayoutComponent } from "./layouts/organization-layout.component";
 import { CollectionsComponent } from "./manage/collections.component";
 import { GroupsComponent } from "./manage/groups.component";
 import { ManageComponent } from "./manage/manage.component";
-import { VaultModule } from "./org-vault/vault.module";
 
 const routes: Routes = [
   {
