+++ apps/web/src/vault/app/org-vault/collection-badge/collection-name-badge.component.html
