+++ apps/web/src/vault/app/org-vault/group-badge/group-name-badge.component.html
