+++ apps/web/src/vault/app/org-vault/vault-filter/vault-filter.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from "@angular/core";
 
-import { VaultFilterService as VaultFilterServiceAbstraction } from "../../../../vault/app/vault/vault-filter/services/abstractions/vault-filter.service";
-import { VaultFilterSharedModule } from "../../../../vault/app/vault/vault-filter/shared/vault-filter-shared.module";
+import { VaultFilterService as VaultFilterServiceAbstraction } from "../../vault/vault-filter/services/abstractions/vault-filter.service";
+import { VaultFilterSharedModule } from "../../vault/vault-filter/shared/vault-filter-shared.module";
 
 import { VaultFilterComponent } from "./vault-filter.component";
 import { VaultFilterService } from "./vault-filter.service";
