+++ apps/web/src/vault/app/org-vault/vault-filter/vault-filter.service.ts
@@ -14,10 +14,10 @@ import { TreeNode } from "@bitwarden/common/models/domain/tree-node";
 import { CipherService } from "@bitwarden/common/vault/abstractions/cipher.service";
 import { FolderService } from "@bitwarden/common/vault/abstractions/folder/folder.service.abstraction";
 
-import { VaultFilterService as BaseVaultFilterService } from "../../../../vault/app/vault/vault-filter/services/vault-filter.service";
-import { CollectionFilter } from "../../../../vault/app/vault/vault-filter/shared/models/vault-filter.type";
-import { CollectionAdminView } from "../../core";
-import { CollectionAdminService } from "../../core/services/collection-admin.service";
+import { CollectionAdminView } from "../../../../app/organizations/core";
+import { CollectionAdminService } from "../../../../app/organizations/core/services/collection-admin.service";
+import { VaultFilterService as BaseVaultFilterService } from "../../vault/vault-filter/services/vault-filter.service";
+import { CollectionFilter } from "../../vault/vault-filter/shared/models/vault-filter.type";
 
 @Injectable()
 export class VaultFilterService extends BaseVaultFilterService implements OnDestroy {
