+++ apps/web/src/vault/app/org-vault/vault-header/vault-header.component.ts
@@ -18,15 +18,15 @@ import {
   SimpleDialogType,
 } from "@bitwarden/components";
 
-import { VaultFilterService } from "../../../../vault/app/vault/vault-filter/services/abstractions/vault-filter.service";
-import { VaultFilter } from "../../../../vault/app/vault/vault-filter/shared/models/vault-filter.model";
-import { CollectionFilter } from "../../../../vault/app/vault/vault-filter/shared/models/vault-filter.type";
-import { CollectionAdminService, CollectionAdminView } from "../../core";
+import { CollectionAdminService, CollectionAdminView } from "../../../../app/organizations/core";
 import {
   CollectionDialogResult,
   CollectionDialogTabType,
   openCollectionDialog,
-} from "../../shared";
+} from "../../../../app/organizations/shared";
+import { VaultFilterService } from "../../vault/vault-filter/services/abstractions/vault-filter.service";
+import { VaultFilter } from "../../vault/vault-filter/shared/models/vault-filter.model";
+import { CollectionFilter } from "../../vault/vault-filter/shared/models/vault-filter.type";
 
 @Component({
   selector: "app-org-vault-header",
